+++ schwarz/fakefs_helpers/fake_fs.py
@@ -1,5 +1,5 @@
 # -*- coding: utf-8 -*-
-# Copyright (c) 2016-2020, 2022 Felix Schwarz
+# Copyright (c) 2016-2020, 2022, 2024 Felix Schwarz
 # The source code contained in this file is licensed under the MIT license or
 # (at your option) the CC0 v1.0.
 # SPDX-License-Identifier: MIT or CC0-1.0
@@ -37,6 +37,12 @@ class FakeFS(object):
 
     def tear_down(self):
         self._stubber.tearDown()
+
+    def __enter__(self):
+        return self
+
+    def __exit__(self, exc_type, exc_value, traceback):
+        self.tear_down()
 
     def __getattr__(self, name):
         if hasattr(self._stubber.fs, name):